+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS25M282XHF2GGGCJKBP0D93"
+        "01KS2RRJ5QFEA6JDPZ147DKGX7"
     )
     r"""The unique identifier of the data source"""