+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS25M2810B2CRS61CPP3C06M"
+        "01KS2RRJ5PTEXE9EHJ646TVGKQ"
     )
     r"""The unique identifier of the data source"""