+++ packages/orq-rc/RELEASES.md
@@ -8069,3 +8069,13 @@ Based on:
 - [python v4.10.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.21 - packages/orq-rc
+
+## 2026-05-20 13:24:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.3 (2.884.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.22 - packages/orq-rc